requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8

# ! move this to dev requirement
factory-boy==2.9.2
